\name{baseball}
\alias{baseball}
\title{Baseball player's perfomance data}
\description{
  Japanese proffesional baseball fielder's performance data in 2014.
}
\usage{
data(baseball)
}
\arguments{
}
\format{
    \enumerate{
        \item player : player's ID
        \item team : team ID
        \item league : 0=central league, 2=Pacific League
        \item AVR : Average
        \item ATbase : 
        \item ATbats : 
        \item HIT : 
        \item HR : Homerun
        \item RBI : Run batted in
        \item walk : base on balls
        \item K : Strike out
        \item salary : Annual salary
        \item salary_log : log transformation of salary
    }
}
\value{
}
\references{http://baseball-freak.com/}
\seealso{}
\examples{
}
\keyword{ }
